End a granted triggered ability when its "for as long as" duration does
`GrantTriggeredAbilityEffect` has always accepted any `Duration`, but
`GameState.grantedTriggeredAbilities` had no gate of its own: `TriggerAbilityResolver`
read the store directly and `EndedDurationExpiryCheck` deliberately skipped it, with a
doc comment saying the first card to grant a trigger "for as long as …" would have to
add it. Makeshift Mannequin is that card, so add it.

Both halves that every other conditional duration gets, now behind one shared
`GrantDurationGate` so they cannot disagree:

- per-read — `TriggerAbilityResolver` skips a grant whose condition is false right now,
  so the drawback goes dark the moment the counter leaves rather than at the next
  state-based-action pass. Mirrors `StateProjector`'s per-frame gate on floating effects.
- one-way latch — `EndedDurationExpiryCheck.pruneEndedGrants` now also prunes
  `grantedTriggeredAbilities`, so re-adding the counter cannot resurrect the ability
  (CR 611.2b).

`GrantedTriggeredAbility` gains a nullable `sourceId` (populated from the effect
context) so the source-keyed gates — `WhileSourceOnBattlefield`, `WhileSourceTapped`,
`WhileSourceAttachedToAffected` — work for a triggered grant too. Every existing grant
into this store uses `EndOfTurn` or `Permanent`, which both gates pass unchanged.
